+++ src/encodings/index.js
@@ -1,0 +1,2 @@
+export { encoding as json } from './json.js';
+export { encoding as jsonxl } from './jsonxl.js';