+++ src/encodings/json.js
@@ -1,0 +1,8 @@
+import { parseChunked } from '@discoveryjs/json-ext';
+
+export const encoding = Object.freeze({
+    name: 'json',
+    test: () => true,
+    streaming: true,
+    decode: parseChunked
+});