+++ src/encodings/jsonxl.js
@@ -1,0 +1,8 @@
+import { decode, isHeaderAcceptable as test } from '../tmp/jsonxl-snapshot9.js';
+
+export const encoding = Object.freeze({
+    name: 'jsonxl/snapshot9',
+    test,
+    streaming: false,
+    decode
+});