+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## next
 
+- Added `gzip` and `deflate` compression support for input data
 - Updated jsonxl
     - Fixed an edge case for signed numbers in the range ±[MAX_SAFE_INTEGER/2 … MAX_SAFE_INTEGER].
     - Removed the limitation on total encoded string length exceeding the maximum string length (~500 MB for V8)